+++ resources.go
@@ -1907,9 +1907,11 @@ func Provider() tfbridge.ProviderInfo {
 							"metric.ts",
 							"metrics.ts",
 							"widgets.ts",
-							"widgetJson.ts",
-							"flowWidgets.ts",
-							"simpleWidgets.ts",
+							"widgets_annotations.ts",
+							"widgets_json.ts",
+							"widgets_flow.ts",
+							"widgets_simple.ts",
+							"widgets_graph.ts",
 						},
 					},
 					"codebuild": {
