+++ sdk/nodejs/tsconfig.json
@@ -124,13 +124,12 @@
         "cloudtrail/trail.ts",
         "cloudwatch/cloudwatchMixins.ts",
         "cloudwatch/dashboard.ts",
-        "cloudwatch/dashboardBody.ts",
+        "cloudwatch/dashboardDescription.ts",
         "cloudwatch/dashboardMixins.ts",
         "cloudwatch/eventPermission.ts",
         "cloudwatch/eventRule.ts",
         "cloudwatch/eventRuleMixins.ts",
         "cloudwatch/eventTarget.ts",
-        "cloudwatch/flowWidgets.ts",
         "cloudwatch/getLogGroup.ts",
         "cloudwatch/index.ts",
         "cloudwatch/logDestination.ts",
@@ -145,9 +144,12 @@
         "cloudwatch/metricAlarm.ts",
         "cloudwatch/metricAlarmMixins.ts",
         "cloudwatch/metrics.ts",
-        "cloudwatch/simpleWidgets.ts",
-        "cloudwatch/widgetJson.ts",
         "cloudwatch/widgets.ts",
+        "cloudwatch/widgets_annotations.ts",
+        "cloudwatch/widgets_flow.ts",
+        "cloudwatch/widgets_graph.ts",
+        "cloudwatch/widgets_json.ts",
+        "cloudwatch/widgets_simple.ts",
         "codebuild/index.ts",
         "codebuild/metrics.ts",
         "codebuild/project.ts",
